move cohort and person property tests to clickhouse
